refactor: move harvestPlot to its own file
