+++ src/game-logic/reducers/index.js
@@ -122,6 +122,7 @@ export * from './fertilizePlot'
 export * from './setSprinkler'
 export * from './processField'
 export * from './setScarecrow'
+export * from './harvestPlot'
 
 ///////////////////////////////////////////////////////////
 //
